Refactor Test Case.

+++ scripts/generate_testcase.py
@@ -1,28 +1,8 @@
 import os
 import datetime
-class TestCase:
-    def __init__(self):
-        self.name = ""
-        self.functions = []
-template ="""
-//
-//  {name}TestCase.swift
-//
-//
-//  Created by lonnie on {date}.
-//
-
-import Foundation
-import XCTest
-@testable import FoundationLib
-
-final class {name}TestCase: XCTestCase {
-    func testSample() {
-    
-    }
-}
-"""
+from models import  *
 if __name__ == "__main__":
+    print(template)
     date = datetime.date.today()
     date_str = "%d/%d/%d"%(date.year, date.month, date.day)
     items = __file__.split("/")
@@ -50,9 +30,23 @@ if __name__ == "__main__":
         if not os.path.exists(dir):
             os.mkdir(dir)
         final_path = dir + "/" + name + "TestCase.swift"
-        content = template.replace("{name}", name).replace("{date}", date_str)
-        with open(final_path, 'w') as f:
-            f.write(content)
+        if os.path.exists(final_path):
+            print("file exists in %s"%(final_path))
+            contents = []
+            with open(final_path, 'r') as f:
+                contents = f.readlines()
+            index = -1
+            for i in range(len(contents) - 1, 0, -1):
+                if contents[i].startswith("}"):
+                    index = i
+            if index != -1:
+                contents = contents[0:index]
+                # TODO! Append necessary functions
+                contents.append("}\n")
+        else:
+            content = template.replace("{name}", name).replace("{date}", date_str)
+            with open(final_path, 'w') as f:
+                f.write(content)
     manifests_path = test_path + "/" + "XCTestManifests.swift"
     current_root = ""
     swift_files = []
@@ -99,7 +93,7 @@ if __name__ == "__main__":
     manifests += "public func allTests() -> [XCTestCaseEntry] {\n\treturn [\n"
     for test_case in test_cases:
         manifests += "\t\ttestCase({name}.allTests),\n".format(name=test_case.name)
-    manifests += """\t]\n}\n#endif\n"""
+    manifests += "\t]\n}\n#endif\n"
     with open(manifests_path, 'w') as f:
         f.write(manifests)
